Changed double to decimal data type for currency amounts

+++ Src/MoneyFox.Application/Statistics/Queries/GetCategorySummary/CategoryOverviewItem.cs
@@ -3,8 +3,8 @@
     public class CategoryOverviewItem
     {
         public string Label { get; set; }
-        public double Value { get; set; }
-        public double Average { get; set; }
-        public double Percentage { get; set; }
+        public decimal Value { get; set; }
+        public decimal Average { get; set; }
+        public decimal Percentage { get; set; }
     }
 }
